move get_major_minor
default triager should not depend on AnsibleVersionIndexer
